requirements.txt
aiofiles
mwparserfromhell
nltk
openai
python-dotenv